+++ actix-service/src/lib.rs
@@ -1,5 +1,6 @@
 use std::cell::RefCell;
 use std::future::Future;
+use std::pin::Pin;
 use std::rc::Rc;
 use std::sync::Arc;
 use std::task::{self, Context, Poll};
@@ -24,6 +25,8 @@ pub use self::fn_service::{factory_fn, factory_fn_cfg, service_fn, service_fn2};
 pub use self::map_config::{map_config, unit_config, MappedConfig};
 pub use self::pipeline::{pipeline, pipeline_factory, Pipeline, PipelineFactory};
 pub use self::transform::{apply, Transform};
+
+pub type BoxFuture<I, E> = Pin<Box<dyn Future<Output = Result<I, E>>>>;
 
 /// An asynchronous function from `Request` to a `Response`.
 pub trait Service {